Move Zero and One to public module
